Fixed and re-done areas of previous commit, plus server-side issues which would/have occured

+++ game/client/enemy.js
@@ -3,20 +3,23 @@
 function Enemy(x, y, name) {
     // Init
     this.characterID = 1;
-    this.playerSprite = game.add.sprite(x, y, "player" + this.characterID);
-    this.playerSprite.width = 32;
-    this.playerSprite.height = 32;
     this.percentage = 0;
     this.name = name;
+    this.lobbyID = 0;
     this.x = x;
     this.y = y;
 
-    // Physics
-    game.physics.arcade.enable(this.playerSprite);
-
-    // Properties
-    //  Body
-    this.playerSprite.body.setSize(64, 64, 0, 0); // 64x64 is default sprite size
+    this.startGame = function() {
+        // Sprite
+        this.playerSprite = game.add.sprite(x, y, "player" + this.characterID);
+        this.playerSprite.width = 32;
+        this.playerSprite.height = 32;
+        // Physics
+        game.physics.arcade.enable(this.playerSprite);
+        // Properties
+        //  Body
+        this.playerSprite.body.setSize(64, 64, 0, 0); // 64x64 is default sprite size
+    };
 
     this.playerUpdate = function() {
         // Stops from falling through the floor
@@ -67,6 +70,8 @@ function Enemy(x, y, name) {
     };
 
     this.remove = function() {
-        this.playerSprite.kill();
+        if (this.playerSprite) {
+            this.playerSprite.kill();
+        }
     };
 }
